# Important optimalization for ascii unicode formatting