@marc1n 2019-07-11
Output for Python 3.7.0b2:
```
Original pprintpp:  0.252066s
Optimized pprintpp: 0.000623s
Standard pprint:    0.001989s
```